refactor: Decompose monolithic helper.py into focused modules
Splits the 3,472-line helper.py into focused modules:

- models/pin.py: Pin dataclass
- models/symbol.py: SymbolDef and SymbolInstance classes
- models/elements.py: Wire, Junction, Label, HierarchicalLabel, PowerSymbol
- models/schematic.py: Schematic class and SnapMode enum
- exceptions.py: Custom exceptions (PinNotFoundError, etc.)
- grid.py: Grid utilities (snap_to_grid, is_on_grid, etc.)
- helpers.py: String utilities and pin alias handling
- logging.py: Verbose logging control
- library.py: Library discovery functions

This improves:
- Code organization and discoverability
- Single responsibility per module
- Easier testing and maintenance
- Clearer import structure

Closes #15
