+++ src/kicad_tools/schematic/models/__init__.py
@@ -1,0 +1,33 @@
+"""
+KiCad Schematic Models
+
+Re-exports all model classes for convenient importing.
+"""
+
+from .elements import (
+    HierarchicalLabel,
+    Junction,
+    Label,
+    PowerSymbol,
+    Wire,
+)
+from .pin import Pin
+from .schematic import Schematic, SnapMode
+from .symbol import SymbolDef, SymbolInstance
+
+__all__ = [
+    # Pin
+    "Pin",
+    # Symbol
+    "SymbolDef",
+    "SymbolInstance",
+    # Elements
+    "Wire",
+    "Junction",
+    "Label",
+    "HierarchicalLabel",
+    "PowerSymbol",
+    # Schematic
+    "Schematic",
+    "SnapMode",
+]
